feat(muk): move and delete hooks to data api folders
ref: #MANAGER-19376
