+++ packages/manager-ui-kit/src/components/filters/index.ts
@@ -1,5 +1,4 @@
 export { FilterAdd } from './filter-add.component';
 export { FilterList } from './filter-list.component';
-export { useColumnFilters } from './useColumnFilters';
 export * from './tags-filter-form.component';
 export type { FilterWithLabel } from './interface';
