+++ packages/manager-ui-kit/src/components/index.ts
@@ -12,8 +12,6 @@ export * from './tiles-input';
 export * from './Link';
 export * from './datagrid/datagrid.component';
 export * from './datagrid/text-cell.component';
-export * from './datagrid/useDatagrid';
-export * from './datagrid/useDatagridSearchParams';
 export * from './datagrid/clipboard-cell.component';
 
 export * from './drawer/Drawer.component';
