+++ packages/manager-ui-kit/src/hooks/data-api/index.ts
@@ -1,1 +1,3 @@
 export * from './ports/use-data-api';
+export { useDatagridSearchParams } from './useDatagridSearchParams';
+export { useColumnFilters } from './useColumnFilters';
