+++ packages/manager-ui-kit/src/hooks/datagrid/useIcebergV6.tsx
@@ -6,7 +6,8 @@ import {
 } from '@ovh-ux/manager-core-api';
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { defaultPageSize } from './index';
-import { DatagridColumn, useColumnFilters, ColumnSort } from '../../components';
+import { DatagridColumn, ColumnSort } from '../../components';
+import { useColumnFilters } from '../data-api/useColumnFilters';
 
 interface IcebergV6Hook<T> {
   queryKey: string[];
