+++ packages/manager-ui-kit/src/hooks/datagrid/useResourcesV6.tsx
@@ -8,7 +8,8 @@ import {
   v6,
 } from '@ovh-ux/manager-core-api';
 import { Query, useQuery } from '@tanstack/react-query';
-import { DatagridColumn, useColumnFilters, ColumnSort } from '../../components';
+import { DatagridColumn, ColumnSort } from '../../components';
+import { useColumnFilters } from '../data-api/useColumnFilters';
 
 export type FetchResultV6<T> = {
   data: T[];
